--------去掉对es的支持和依赖，默认去掉对hive的包依赖，maven依赖关系优化，页面位置调整，增加build.bat的支持，支持jar直接启动，去掉历史遗留的无用前端代码，依赖的maven包升级--------

+++ zyplayer-doc-manage/src/main/java/com/zyplayer/doc/manage/web/manage/UserInfoController.java
@@ -23,7 +23,7 @@ import com.zyplayer.doc.manage.web.manage.vo.AuthInfoVo;
 import com.zyplayer.doc.manage.web.manage.vo.UserAuthVo;
 import com.zyplayer.doc.manage.web.manage.vo.UserInfoAuthVo;
 import org.apache.commons.lang.StringUtils;
-import org.dozer.Mapper;
+import com.github.dozermapper.core.Mapper;
 import org.springframework.util.DigestUtils;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -120,7 +120,7 @@ public class UserInfoController {
 		QueryWrapper<UserInfo> queryWrapper = new QueryWrapper<>();
 		queryWrapper.eq("user_no", userInfo.getUserNo());
 		queryWrapper.ne(userId > 0, "id", userInfo.getId());
-		int count = userInfoService.count(queryWrapper);
+		long count = userInfoService.count(queryWrapper);
 		if (count > 0) {
 			return DocResponseJson.warn("改用户账号已存在");
 		}
